Fix arrow styles for ListSelectionView
Add support to construct (and initialize) a ListSelectionView from GuiTools, to work around a ControlsFX issue that can make the arrows resistant to changes in style via css.

Add `GuiTools.ensureDuplicatableGlyph(glyph)` method to make this more generally useful, extracting the logic from `IconFactory`.
